박스) 동기화 및 변수 수정

+++ Assets/_StrangeMuseum/kj/Scripts/Interface/IUsableItem.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IUsableItem
 {
-    public void UseServerRpc(ulong id);
+    public void UseServerRpc(uint id);
 
     public ItemData.ItemList GetItemList();
     public ItemData.ItemUseType GetItemType();
